Add setting to disable voice activity indicators
Per feedback [1], the test for an invisible party list doesn't properly
match up with whether the party list is actually visible. If it's hidden
beacuse you're in solo mode and you have the "hide party list when solo"
setting enabled, it works fine – but if you hid it through the HUD
layout settings, it doesn't.

There's no obvious way to determine whether the party list is *actually*
visible, so we add a setting to manually hide the indicators instead.

[1]: https://discord.com/channels/581875019861328007/892131028188151808/1144732215150915604