Never skip figuring out which indicators to draw
The plugin knows who's talking at all times. It knows this because it
knows who isn't talking. By subtracting who is talking from who isn't,
or who isn't talking from who is (whichever is greater), it obtains a
difference, or deviation.

...as a side-effect of drawing the voice activity indicators, we figure
out which users we've drawn indicators for. This information is critical
to correctly populate the list of speaking Discord users who aren't in
the user's party!

Regardless of the state of the new "show voice activity indicators"
setting, always go through the motions of working out who's talking.